Merge Named Pipe system

+++ Bolsa/companyManager.h
@@ -11,10 +11,10 @@ namespace CompanyManager {
 	} OPERATION;
 
 	void config(BOLSA& servidor);
-	void addCompany(BOLSA &servidor, std::TSTRING name, DWORD numOfStock, DWORD pricePerStock);
+	void addCompany(BOLSA &servidor, std::TSTRING name, DWORD numOfStock, float pricePerStock);
 	void listCompanies(BOLSA &servidor);
 	COMPANY* getCompany(COMPANY_LIST& companyList, std::TSTRING name);
-	void updateStock(BOLSA& servidor, std::TSTRING name, DWORD pricePerStock);
+	void updateStock(BOLSA& servidor, std::TSTRING name, float pricePerStock);
 	void updateStock(NOTIFY_DATA& notifyData, COMPANY& company, OPERATION opType);
 	void pauseCompaniesOps(BOLSA& servidor, int time);
 	DWORD WINAPI timerRoutine(LPVOID lpParam);
